Add files via upload
Sequence SM adds functionality of event checker for joystick input. This Event checker uses the Touch sensor (button on joystick was not working properly on my end). Sequence is debugged and working properly. Need to add Postings to GAME and OLED SM.

Added a module function to compare the sequence direction to the user input.

Need to update timers in ES_Configure